% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_source.R
\name{add_source}
\alias{add_source}
\title{Add source to plotly chart}
\usage{
add_source(src, lnk = "", y = -0.1)
}
\arguments{
\item{src}{Source text to show}

\item{lnk}{Link text (if blank the source text will not be a link)}

\item{y}{Where on the y axis (paper) the link will appear defaults to bottom right corner (-.1)}
}
\value{
List to add to annotations object withing layout list of plotly chart
}
\description{
Add source to plotly chart
}
\examples{
add_source(src = "London Datastore", lnk = "data.london.gov.uk")

}
