% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_from_data.R
\name{chart_from_data}
\alias{chart_from_data}
\title{Chart from data - produces chart for single indicator}
\usage{
chart_from_data(dtst, txtfun = NULL)
}
\arguments{
\item{dtst}{Character string that is primary key for indicator}
}
\value{
Plotly chart corresponding to indicator key given
}
\description{
Chart from data - produces chart for single indicator
}
\examples{
chart_from_data("unempl")

}
