% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_from_data.R
\name{data_from_csv}
\alias{data_from_csv}
\title{get data for single indicator from csv files}
\usage{
data_from_csv(x, na.rm = TRUE)

data_from_csv(x, na.rm = TRUE)
}
\arguments{
\item{dr}{Path where csv files are kept}

\item{dtst}{Character string of dataset primary key}

\item{fl}{Path location of RData file}
}
\description{
get data for single indicator from csv files

get data for single indicator from RData files
}
