% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_from_data.R
\name{data_from_sqlite}
\alias{data_from_sqlite}
\title{get data for single indicator from sqlite database}
\usage{
data_from_sqlite(dtst, fl)
}
\arguments{
\item{dtst}{Character string of dataset primary key}

\item{fl}{Path location of sqlite file}
}
\description{
get data for single indicator from sqlite database
}
