% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsh_bar.R
\name{dsh_bar}
\alias{dsh_bar}
\title{Produce a single bar chart in plotly}
\usage{
dsh_bar(
  dat,
  ttl,
  yxttl = "",
  xxttl = "",
  src = "",
  lnk = "",
  tckfmt = "",
  hv = "\%{y}"
)
}
\arguments{
\item{dat}{data frame with correct column headings}

\item{ttl}{Chart title (character)}

\item{yxttl}{Character string of title for yaxis}

\item{xxttl}{Character string of title for xaxis}

\item{src}{Character string of source of data}

\item{tckfmt}{Tick format (eg \% for percentage)}

\item{hv}{hovertemplate}

\item{link}{Character string of link of source}
}
\value{
Single bar plotly chart object (one category)
}
\description{
Produce a single bar chart in plotly
}
\examples{
chart_from_data("unemp")

}
