% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsh_mbar.R
\name{dsh_mbar}
\alias{dsh_mbar}
\title{Produce a multi bar chart in plotly}
\usage{
dsh_mbar(
  dat,
  ttl,
  yxttl = "",
  xxttl = "",
  src = "",
  lnk = "",
  tckfmt = "",
  hv = "\%{y}",
  stack = FALSE
)
}
\arguments{
\item{dat}{data frame with correct column headings}

\item{ttl}{Chart title (character)}

\item{yxttl}{Character string of title for yaxis}

\item{xxttl}{Character string of title for xaxis}

\item{src}{Character string of source of data}

\item{tckfmt}{Tick format (eg \% for percentage)}

\item{hv}{hovertemplate}

\item{link}{Character string of link of source}
}
\value{
Multi-bar plotly chart object (stacked or dodged - more than one category)
}
\description{
Produce a multi bar chart in plotly
}
\examples{
chart_from_data("unemp")

}
