% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsh_sub_region_line.R
\name{dsh_sub_region_line}
\alias{dsh_sub_region_line}
\title{Produce a multi bar chart in plotly}
\usage{
dsh_sub_region_line(
  dat,
  ttl,
  xlb = "",
  ylb = "",
  src = "",
  lnk = "",
  hv = "\%{y}"
)
}
\arguments{
\item{dat}{data frame with correct column headings}

\item{ttl}{Chart title (character)}

\item{xlb}{Character string of title for xaxis}

\item{ylb}{Character string of title for yaxis}

\item{src}{Character string of source of data}

\item{hv}{hovertemplate}

\item{link}{Character string of link of source}
}
\value{
Multi-bar plotly chart object (stacked or dodged - more than one category)
}
\description{
Produce a multi bar chart in plotly
}
\examples{
chart_from_data("unemp")

}
