% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dashboard.R
\name{generate_dashboard}
\alias{generate_dashboard}
\title{Generates dashboard from RMarkdown file saves it to repo}
\usage{
generate_dashboard()
}
\value{
TRUE if it worked
}
\description{
Generates dashboard from RMarkdown file saves it to repo
}
\examples{
generate_dashboard()


}
