% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_from_data.R
\name{get_data_for_dataset}
\alias{get_data_for_dataset}
\title{get data for single indicator}
\usage{
get_data_for_dataset(dtst)
}
\arguments{
\item{dtst}{Character string of dataset primary key}
}
\description{
get data for single indicator
}
