% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_dashboard.R
\name{publish_dashboard}
\alias{publish_dashboard}
\title{Publish dashboard to gitea repo}
\usage{
publish_dashboard(msg)
}
\arguments{
\item{msg}{Message to add to the commit}
}
\value{
TRUE if it worked
}
\description{
Publish dashboard to gitea repo
}
\examples{
msg <- "Commit message" 
publish_dashboard(msg)


}
