% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dashboard.R
\name{generate_dashboard}
\alias{generate_dashboard}
\title{Generates dashboard RMarkdown file}
\usage{
generate_dashboard(drct, knitit = FALSE)
}
\arguments{
\item{drct}{Path to output file}

\item{knitit}{Logical - if TRUE, knits the generated RMarkdown file}
}
\value{
TRUE if it worked
}
\description{
Generates dashboard RMarkdown file
}
\examples{

path_to_output_file <- "path/to/output/dash.Rmd"

generate_dashboard(path_to_output_file, TRUE)

## generates 

}
